Moved the file-based interface to its own python file.
